Optimize Pokemon 'L' script

+++ catscripts/Government/Countries/Cuba/Cuba.sh
@@ -1,14 +1,26 @@
 #!/bin/bash
 
-egrep -i '\bCuba|Fidel Castro|Raul Castro|Havana' newpages.txt >> Cuba.txt
+KEYWORDS_CUBA="\bCuba|Fidel Castro|Raul Castro|Havana"
+KEYWORDS_CUBA_EXCLUDE="Cuba(| )Gooding"
+KEYWORDS_CUBA_ALL="$KEYWORDS_CUBA"
 
-CUBA=`stat --print=%s Cuba.txt`
-
-if [ $CUBA -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Cuba.txt"
-  export CATNAME="Cuba"
-  $CATEGORIZE
-fi
 
-rm Cuba.txt
+  debug_start "Cuba"
+
+  CUBA="$(egrep -i "$KEYWORDS_CUBA" newpages.txt | egrep -iv "$KEYWORDS_CUBA_EXCLUDE")"
+
+  if [ "$CUBA" != "" ];
+  then
+    printf "%s" "$CUBA" > Cuba.txt
+    export CATFILE="Cuba.txt"
+    export CATNAME="Cuba"
+    $CATEGORIZE
+    rm Cuba.txt
+    unset CUBA
+  fi
+
+  debug_end "Cuba"
+
+fi
